+++ Editor/ZapparFaceMeshTargetEditor.cs
@@ -14,9 +14,9 @@ namespace Zappar.Editor
         private void OnEnable()
         {
             m_target = (ZapparFaceMeshTarget)target;
-            if (m_target.FaceTrackingTarget == null)
+            if (m_target.FaceTrackingAnchor == null)
             {
-                Debug.Log("Assign Face tracking target for this face mesh");
+                Debug.Log("Assign Face tracking anchor for this face mesh");
                 return; 
             }
             m_usingFullHead = m_target.UseDefaultFullHead;